requirements.txt
pandas
matplotlib
reportlab
xlsxwriter
psycopg2-binary